requirements.txt
boto3==1.18.44
botocore==1.21.44
jmespath==0.10.0
numpy==1.21.2
opencv-python==4.4.0.42
Pillow==8.3.2
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
